tests: finish converting to jest and all runnin successfully again
fixed issue of tests hanging and not sending any exit signal,
previously incorrectly causing yarn to believe tests succeded even
when they did not.

the issue was about puppeteer chromium browsers being still open,
and hence the test runners not ending the tests. I could have chosen
to force exit via test runner (after tests), but instead for now I
cleaned up the browsers after the tests, which is a bit tricky given
the tests for parallel runs and the unstableKeepBrowserOpen option.
