+++ test/basic.test.js
@@ -8,16 +8,13 @@ import csstree from 'css-tree'
 import normaliseCss from './util/normaliseCss'
 
 function staticServerFileUrl (file) {
-  return 'file://' + path.join(__dirname, 'static-server', file)
+  return 'file://' + path.join(process.env.PWD, 'test', 'static-server', file)
 }
 
 describe('basic tests of penthouse functionality', () => {
   var page1FileUrl = staticServerFileUrl('page1.html')
-  var page1cssPath = path.join(__dirname, 'static-server', 'page1.css')
+  var page1cssPath = path.join(process.env.PWD, 'test', 'static-server', 'page1.css')
   var originalCss = read(page1cssPath).toString()
-
-  // some of these tests take longer than default timeout
-  this.timeout(10000)
 
   it('should return css', () => {
     return penthouse({
@@ -25,7 +22,7 @@ describe('basic tests of penthouse functionality', () => {
       css: page1cssPath
     })
       .then(result => {
-        expect(result).to.have.length.greaterThan(0)
+        expect(result.length).toBeGreaterThan(0)
       })
   })
 
@@ -65,30 +62,30 @@ describe('basic tests of penthouse functionality', () => {
   it('should not crash on invalid css', () => {
     return penthouse({
       url: page1FileUrl,
-      css: path.join(__dirname, 'static-server', 'invalid.css')
+      css: path.join(process.env.PWD, 'test', 'static-server', 'invalid.css')
     })
       .then(result => {
         if (result.length === 0) {
           throw new Error('length should be > 0')
         }
-        expect(result).to.have.length.greaterThan(0)
+        expect(result.length).toBeGreaterThan(0)
       })
   })
 
   it('should not crash on invalid media query', () => {
     return penthouse({
       url: page1FileUrl,
-      css: path.join(__dirname, 'static-server', 'invalid-media.css')
+      css: path.join(process.env.PWD, 'test', 'static-server', 'invalid-media.css')
     })
       .then(result => {
-        expect(result).to.have.length.greaterThan(0)
+        expect(result.length).toBeGreaterThan(0)
       })
   })
 
   it('should crash with errors in strict mode on invalid css', done => {
     penthouse({
       url: page1FileUrl,
-      css: path.join(__dirname, 'static-server', 'invalid.css'),
+      css: path.join(process.env.PWD, 'test', 'static-server', 'invalid.css'),
       strict: true
     })
       .then(() => done(new Error('Did not get error')))
@@ -98,10 +95,10 @@ describe('basic tests of penthouse functionality', () => {
   it('should not crash or hang on special chars', () => {
     return penthouse({
       url: page1FileUrl,
-      css: path.join(__dirname, 'static-server', 'special-chars.css')
+      css: path.join(process.env.PWD, 'test', 'static-server', 'special-chars.css')
     })
       .then(result => {
-        expect(result).to.have.length.greaterThan(0)
+        expect(result.length).toBeGreaterThan(0)
       })
   })
